wines_backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Django==3.1.4
django-cors-headers==3.6.0
django-environ==0.4.5
django-filter==2.4.0
django-parler==2.2
django-parler-rest==2.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
drf-nested-multipart-parser==0.0.2
dynaconf==3.1.2
environ==1.0
hvac==0.10.6
idna==2.10
loguru==0.5.3
MarkupSafe==1.1.1
pathlib==1.0.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pymaging==0.1
pymaging-png==0.1
pytz==2020.5
qrcode==6.1
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
validators==0.18.2
